pyproject.toml
[project]
name = "lenticular-printing"
version = "0.1.0"
description = "Programmatic lenticular print interlacing — configurable lens and printer profiles"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0",
    "numpy>=1.24",
]

[project.scripts]
lenticular = "lenticular.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lenticular"]
